moved extension ui interfaces to own file and linting issues
