+++ spiffworkflow-backend/src/spiffworkflow_backend/scripts/set_user_properties.py
@@ -1,11 +1,11 @@
 from typing import Any
-from spiffworkflow_backend.models.db import db
 
-from flask import current_app
 from flask import g
+from spiffworkflow_backend.models.db import db
 from spiffworkflow_backend.models.script_attributes_context import ScriptAttributesContext
 from spiffworkflow_backend.models.user_property import UserPropertyModel
-from spiffworkflow_backend.scripts.script import InvalidArgsGivenToScript, Script
+from spiffworkflow_backend.scripts.script import InvalidArgsGivenToScriptError
+from spiffworkflow_backend.scripts.script import Script
 
 
 class SetUserProperties(Script):
@@ -20,7 +20,9 @@ class SetUserProperties(Script):
     def run(self, script_attributes_context: ScriptAttributesContext, *args: Any, **kwargs: Any) -> Any:
         properties = args[0]
         if not isinstance(properties, dict):
-            raise InvalidArgsGivenToScript(f"Args to set_user_properties must be a dict. '{properties}' is invalid.")
+            raise InvalidArgsGivenToScriptError(
+                f"Args to set_user_properties must be a dict. '{properties}' is invalid."
+            )
         # consider using engine-specific insert or update metaphor in future: https://stackoverflow.com/a/68431412/6090676
         for property_key, property_value in properties.items():
             user_property = UserPropertyModel.query.filter_by(user_id=g.user.id, key=property_key).first()
